requirements.txt
GSAS-II @ git+https://github.com/AdvancedPhotonSource/GSAS-II.git
streamlit
pandas
pyyaml
--extra-index-url https://download.pytorch.org/whl/cpu
torch
scikit-learn
pymatgen
plotly
matplotlib
Pillow
gdown
h5py
imageio
pycifrw
xmltodict
psutil
# Core API & Gateway
fastapi
uvicorn[standard]
python-multipart
httpx
